Haste

Description

Haste's companion to the Prelude.

Basic utility functions

data JSString :: *

data JSAny :: *

type URL = JSString

class Monad m => MonadIO m where

Monads in which IO computations may be embedded. Any monad built by applying a sequence of monad transformers to the IO monad will be an instance of this class.

Instances should satisfy the following laws, which state that liftIO is a transformer of monads:

liftIO . return = return

liftIO (m >>= f) = liftIO m >>= (liftIO . f)

Methods

liftIO :: IO a -> m a

Lift a computation from the IO monad.

alert :: MonadIO m => JSString -> m ()

JavaScript alert() function.

prompt :: MonadIO m => JSString -> m JSString

JavaScript prompt() function.

eval :: MonadIO m => JSString -> m JSString

JavaScript eval() function.

writeLog :: MonadIO m => JSString -> m ()

JavaScript console.log().

catJSStr :: JSString -> [JSString] -> JSString

fromJSStr :: JSString -> String

Location handling

onHashChange :: MonadEvent m => (JSString -> JSString -> m ()) -> m ()

Register a callback to be run whenever the URL hash changes. The first and second argument of the callback are the old and new and hash respectively.

setHash :: MonadIO m => JSString -> m ()

Set the hash part of the current URL.

getHash :: MonadIO m => m JSString

Read the hash part of the current URL.

getLocationHref :: MonadIO m => m URL

Get the current complete location URL.

setLocationHref :: MonadIO m => URL -> m ()

Set the location URL.

getLocationHostName :: MonadIO m => m JSString

Get the current location host name.

getLocationPort :: MonadIO m => m Int

Get the current location port.

Timers

data Timer

Timer handle.

data Interval

Interval and repeat for timers.

Constructors

Once !Int
Fire once, in n milliseconds.

Repeat !Int
Fire every n milliseconds.

setTimer

Arguments

:: MonadEvent m

=> Interval
Milliseconds until timer fires.

-> m ()
Function to call when timer fires.

-> m Timer
Timer handle for interacting with the timer.

Set a timer.

stopTimer :: MonadIO m => Timer -> m ()

Stop a timer.

Fast conversions for JS-native types

class JSType a where

Methods

toJSString :: a -> JSString

fromJSString :: JSString -> Maybe a

class JSNum a where

Methods

toNumber :: a -> Double

fromNumber :: Double -> a

toString :: JSType a => a -> String

fromString :: JSType a => String -> Maybe a

convert :: (JSNum a, JSNum b) => a -> b

Reflection

getProgramId :: MonadIO m => m JSString

Get the value of the __haste_prog_id variable. Unless programmatically changed, this variable contains the SHA3-256 hash of the currently executing Haste program.

getProgramJS :: MonadIO m => m (Either URL JSString)

Get the complete JavaScript source code of the currently executing Haste program. On IE, this requires that the program's identifier, as returned by getProgramId, has not been tampered with.