Haste.Ajax

Description

XMLHttpRequest support. IE9 and older are not supported.

Documentation

data Method a where

HTTP method to use for request. POST requests take an (optionally empty) JSString representing data to POST.

Constructors

GET :: Method ()

DELETE :: Method ()

POST :: a -> Method a

PUT :: a -> Method a

class (ToAny a, FromAny a) => AjaxData a where

Methods

mimeType :: Proxy a -> JSString

The MIME type of the data represented by this type.

responseType :: Proxy a -> JSString

The XHR response type that corresponds to this type.

data AjaxError

An error which occurred during an AJAX request. Might be either a network error (denied by CSP, host unreachable etc.) or an HTTP error, with status code and description.

Constructors

NetworkError

HttpError Int JSString

ajax

Arguments

:: (MonadConc m, AjaxData post, AjaxData resp)

=> Method post
HTTP method to use for request. post is always () for GET requests.

-> URL
URL to request.

-> m (Either AjaxError resp)

Make an AJAX request. The return value is either the body of the requested document, or an error.

ajaxWithMime

Arguments

:: (MonadConc m, AjaxData post, AjaxData resp)

=> JSString
MIME type of any POST data. Decided by the AjaxData instance for post if the empty string is given. Only relevant to POST requests.

-> Method post
HTTP method to use for request.

-> URL
URL to request.

-> m (Either AjaxError resp)

Like ajax, but accepts a custom MIME type for POST data. Use this with the appropriate MIME type and POST data if you want to send, for instance, form data.