Haste.Binary

Description

Handling of Javascript-native binary blobs.

Generics borrowed from the binary package by Lennart Kolmodin (released under BSD3)

High level binary API

class MonadIO m => MonadConc m

Any monad which supports concurrency.

Minimal complete definition

liftCIO, fork

class Binary a where

Somewhat efficient serializationdeserialization tofrom binary Blobs. The layout of the binaries producedread by getput and encode/decode may change between versions. If you need a stable binary format, you should make your own using the primitives in Haste.Binary.Get/Put.

Minimal complete definition

Nothing

Methods

get :: Get a

put :: a -> Put

data Blob

A JavaScript Blob on the client, a ByteString on the server.

data BlobData

In a browser context, BlobData is essentially a DataView, with an accompanying offset and length for fast slicing. In a server context, it is simply a ByteString.

encode :: Binary a => a -> Blob

Encode any serializable data into a Blob.

decode :: Binary a => BlobData -> Either JSString a

Decode any deserializable data from a BlobData.

decodeBlob :: (MonadConc m, Binary a) => Blob -> m (Either JSString a)

Decode a Blob into some deserializable value, inconveniently locked up inside the CIO monad (or any other concurrent monad) due to the somewhat special way JavaScript uses to deal with binary data.

type Put = PutM ()

data PutM a

putWord8 :: Word8 -> Put

putWord16le :: Word16 -> Put

putWord32le :: Word32 -> Put

putInt8 :: Int8 -> Put

putInt16le :: Int16 -> Put

putInt32le :: Int32 -> Put

putFloat32le :: Float -> Put

putFloat64le :: Double -> Put

putBlob :: Blob -> Put

putJSString :: JSString -> Put

Serialize a JSString as UTF-16 (somewhat) efficiently.

runPut :: Put -> Blob

data Get a

getWord8 :: Get Word8

getWord16le :: Get Word16

getWord32le :: Get Word32

getInt8 :: Get Int8

getInt16le :: Get Int16

getInt32le :: Get Int32

getFloat32le :: Get Float

getFloat64le :: Get Double

getBytes :: Int -> Get BlobData

getJSString :: Int -> Get JSString

skip :: Int -> Get ()

Skip n bytes of input.

runGet :: Get a -> BlobData -> Either JSString a

data Word8 :: *

8-bit unsigned integer type

data Word16 :: *

16-bit unsigned integer type

data Word32 :: *

32-bit unsigned integer type

data Word64 :: *

64-bit unsigned integer type

data Int8 :: *

8-bit signed integer type

data Int16 :: *

16-bit signed integer type

data Int32 :: *

32-bit signed integer type

data Int64 :: *

64-bit signed integer type

Working with raw blobs

class Ord a => Ix a

The Ix class is used to map a contiguous subrange of values in a type onto integers. It is used primarily for array indexing (see the array package).

The first argument (l,u) of each of these operations is a pair specifying the lower and upper bounds of a contiguous subrange of values.

An implementation is entitled to assume the following laws about these operations:

inRange (l,u) i == elem i (range (l,u))

range (l,u) !! index (l,u) i == i, when inRange (l,u) i

map (index (l,u)) (range (l,u))) == [0..rangeSize (l,u)-1]

rangeSize (l,u) == length (range (l,u))

Minimal complete definition

range, (index | unsafeIndex), inRange

class (IArray UArray a, MArray IOUArray a IO) => ArrView a

Minimal complete definition

arrView

getBlobData :: MonadConc m => Blob -> m BlobData

Retrieve the raw data from a blob.

getBlobText' :: MonadConc m => Blob -> m JSString

Interpret a blob as UTF-8 text, as a JSString.

getBlobText :: MonadConc m => Blob -> m String

Interpret a blob as UTF-8 text.

blobSize :: Blob -> Int

The size, in bytes, of the contents of the given blob.

blobDataSize :: BlobData -> Int

The size, in bytes, of the contents of the given blob data.

toByteString :: BlobData -> ByteString

Convert a BlobData to a ByteString. Only usable server-side.

fromByteString :: ByteString -> BlobData

Convert a ByteString to a BlobData. Only usable server-side.

toBlob :: BlobData -> Blob

Convert a piece of BlobData back into a Blob.

strToBlob :: JSString -> Blob

Create a Blob from a JSString.

toUArray :: (Ix i, ArrView e) => BlobData -> UArray i e

Convert a BlobData to an unboxed array; client-side only.

fromUArray :: (Ix i, ArrView e) => UArray i e -> Blob

Convert a an unboxed array into a Blob; client-side only.