Haste.Concurrent

Description

Concurrency for Haste. Includes MVars, forking, Ajax and more.

Documentation

data MVar a

data CIO a

Concurrent IO monad. The normal IO monad does not have concurrency capabilities with Haste. This monad is basically IO plus concurrency.

class MonadIO m => MonadConc m where

Any monad which supports concurrency.

Methods

liftCIO :: CIO a -> m a

fork :: m () -> m ()

forkIO :: CIO () -> CIO ()

Spawn a new thread.

forkMany :: [CIO ()] -> CIO ()

Spawn several threads at once.

newMVar :: MonadIO m => a -> m (MVar a)

Create a new MVar with an initial value.

newEmptyMVar :: MonadIO m => m (MVar a)

Create a new empty MVar.

takeMVar :: MonadConc m => MVar a -> m a

Read an MVar. Blocks if the MVar is empty. Only the first writer in the write queue, if any, is woken.

putMVar :: MonadConc m => MVar a -> a -> m ()

Write an MVar. Blocks if the MVar is already full. Only the first reader in the read queue, if any, is woken.

withMVarIO :: MonadConc m => MVar a -> (a -> IO b) -> m b

Perform an IO action over an MVar.

modifyMVarIO :: MonadConc m => MVar a -> (a -> IO (a, b)) -> m b

Perform an IO action over an MVar, then write the MVar back.

readMVar :: MonadConc m => MVar a -> m a

Read an MVar then put it back. As Javascript is single threaded, this function is atomic. If this ever changes, this function will only be atomic as long as no other thread attempts to write to the MVar.

concurrent :: CIO () -> IO ()

Run a concurrent computation. Two different concurrent computations may share MVars; if this is the case, then a call to concurrent may return before all the threads it spawned finish executing.

liftIO :: MonadIO m => forall a. IO a -> m a

Lift a computation from the IO monad.

tryTakeMVar :: MonadConc m => MVar a -> m (Maybe a)

Try to take a value from an MVar, but return Nothing if it is empty.

tryPutMVar :: MonadConc m => MVar a -> a -> m Bool

Try to put a value into an MVar, returning False if the MVar is already full.

wait :: Int -> CIO ()

Wait for n milliseconds.

withResult :: CIO a -> (a -> IO ()) -> IO ()

When the given concurrent computation is done, pass the result to the given callback.