Haste.Crypto

Documentation

class MonadIO m => MonadConc m where

Any monad which supports concurrency.

Methods

liftCIO :: CIO a -> m a

fork :: m () -> m ()

class Monad m => MonadIO m where

Monads in which IO computations may be embedded. Any monad built by applying a sequence of monad transformers to the IO monad will be an instance of this class.

Instances should satisfy the following laws, which state that liftIO is a transformer of monads:

liftIO . return = return

liftIO (m >>= f) = liftIO m >>= (liftIO . f)

Methods

liftIO :: IO a -> m a

Lift a computation from the IO monad.

class Binary a where

Somewhat efficient serializationdeserialization tofrom binary Blobs. The layout of the binaries producedread by getput and encode/decode may change between versions. If you need a stable binary format, you should make your own using the primitives in Haste.Binary.Get/Put.

Minimal complete definition

Nothing

Methods

get :: Get a

put :: a -> Put

class (IArray UArray a, MArray IOUArray a IO) => ArrView a

Minimal complete definition

arrView

data Word8 :: *

8-bit unsigned integer type

data Word16 :: *

16-bit unsigned integer type

data Word32 :: *

32-bit unsigned integer type

data Word64 :: *

64-bit unsigned integer type

Crypto-strength random number generation

class CryptoRandom a where

Minimal complete definition

cryptoRandom

Methods

cryptoRandom :: MonadIO m => m a

Generate an element of type a using the browser's crypto strength random bit generator.

cryptoRandoms :: MonadIO m => Word32 -> m [a]

Generate n random elements of type a.

cryptoRandomSalt :: MonadIO m => Word32 -> m Salt

Generate a random salt of the given length.

cryptoRandomUArray :: (MonadIO m, ArrView a) => Word32 -> m (UArray Word32 a)

Generate a random immutable array of n elements.

cryptoRandomIOUArray :: (MonadIO m, ArrView a) => Word32 -> m (IOUArray Word32 a)

Generate a random mutable array of n elements.

SubtleCrypto API. Presently only supports symmetric encryption, due to

data Cipher

Constructors

AES CipherMode KeyLength

data CipherMode

Denotes a block cipher mode. As only CBC and GCM are universally supported by browsers, these are presently the only modes exported by Haste. If unsure, always use GCM.

Constructors

CBC

GCM

data KeyLength

Constructors

Bits128

Bits256

data SymmetricKey

A key for a symmetric cipher.

newtype IV

Constructors

IV

Fields

ivBytes :: UArray Word32 Word8

newtype Salt

Constructors

Salt

Fields

saltBytes :: UArray Word32 Word8

type Password = JSString

generateKey :: MonadConc m => Cipher -> m SymmetricKey

Generate a key for the given cipher.

generateIV :: MonadIO m => Cipher -> m IV

Generate an byte initialization vector of appropriate length for the given cipher.

keyFromBytes :: MonadConc m => Cipher -> UArray Word32 Word8 -> m (Either JSString SymmetricKey)

Import a key from an unboxed byte array.

getKeyBytes :: MonadConc m => SymmetricKey -> m (UArray Word32 Word8)

Get the specified key as an unboxed byte array. Only the bytes of the key itself are returned; exported keys do not include information about which cipher they are for.

deriveKey :: MonadConc m => Cipher -> Int -> Salt -> Password -> m (Maybe SymmetricKey)

Derive a key from the given password and salt, using the PBKDF2 key derivation over SHA-256, with the given number of iterations.

One million iterations take roughly half a second on a modern laptop. Returns Nothing on browsers where the operation is not supported. Presently not supported by IE/Edge.

keyCipher :: SymmetricKey -> Cipher

encrypt :: (Binary a, MonadConc m) => SymmetricKey -> a -> m Blob

Encrypt and encrypt a Binary value using a random initialization vector. The IV will be prepended to the message; when using GCM mode, the IV will be 12 bytes, otherwise it will match the length of the key.

An unboxed array can be obtained from the resulting Blob:

toUArray $ getBlobData blob

decrypt :: (Binary a, MonadConc m) => SymmetricKey -> Blob -> m (Either JSString a)

Decrypt and decode a Binary value from a Blob. The initialization vector will be read from the beginning of the message; see encrypt for information about IV lengths. A blob can be obtained from an unboxed array using fromUArray.

encryptWithIV :: (Binary a, MonadConc m) => IV -> SymmetricKey -> a -> m Blob

Encrypt a message using a custom initialization vector. The IV will not be prepended to the output.

decryptWithIV :: (Binary a, MonadConc m) => IV -> SymmetricKey -> Blob -> m (Either JSString a)

Decrypt a message using a custom initialization vector.

encryptUArray :: (Ix i, ArrView e, MonadConc m) => SymmetricKey -> IV -> UArray i e -> m (UArray Word32 Word8)

Encrypt a message using the given cipher, key and initialization vector.

decryptUArray :: (Ix i, ArrView e, MonadConc m) => SymmetricKey -> IV -> UArray i e -> m (Either JSString (UArray Word32 Word8))

Encrypt a message using the given cipher, key and initialization vector.