Haste.Foreign

Description

High level JavaScript foreign interface.

Conversion to/from JSAny

class (IArray UArray a, MArray IOUArray a IO) => ArrView a

Minimal complete definition

arrView

class ToAny a where

Minimal complete definition

Nothing

Methods

toAny :: a -> JSAny

listToAny :: [a] -> JSAny

class FromAny a where

Minimal complete definition

Nothing

Methods

fromAny :: JSAny -> IO a

listFromAny :: JSAny -> IO [a]

data JSAny :: *

data Opaque a :: * -> *

toOpaque :: a -> Opaque a

fromOpaque :: Opaque a -> a

nullValue :: JSAny

toObject :: [(JSString, JSAny)] -> JSAny

has :: JSAny -> JSString -> IO Bool

get :: FromAny a => JSAny -> JSString -> IO a

index :: FromAny a => JSAny -> Int -> IO a

isUndefined :: JSAny -> Bool

getMaybe :: FromAny a => JSAny -> JSString -> IO (Maybe a)

Read a member from a JS object. Succeeds if the member exists.

hasAll :: JSAny -> [JSString] -> IO Bool

Checks if a JS object has a list of members. Succeeds if the JS object has every member given in the list.

lookupAny :: JSAny -> JSString -> IO (Maybe JSAny)

Looks up an object by a .-separated string. Succeeds if the lowest member exists.

Usage example:

>>> {'child': {'childrer': {'childerest': "I am very much a child."}}}

Given the JS Object above, we can access the deeply nested object, childerest, by lookupAny as in the below example

>>> lookupAny jsObject "child.childrer.childerest"

data JSException :: *

Constructors

JSException JSString

Importing and exporting JavaScript functions

class FFI a

class JSFunc a

Minimal complete definition

mkJSFunc, arity

ffi :: FFI a => JSString -> a

constant :: FromAny a => JSString -> a

export :: ToAny a => JSString -> a -> IO ()

safe_ffi :: FFI a => StaticPtr JSString -> a

data StaticPtr a :: * -> *

A reference to a value of type a.

withLibraries :: [JSString] -> IO () -> IO ()

Wait for the given libraries to load before executing the given computation. Note that this function returns immediately, BEFORE the given libraries are loaded.