Haste.LocalStorage

Description

Basic bindings to HTML5 WebStorage.

Documentation

setItem :: Serialize a => JSString -> a -> IO ()

Locally store a serializable value.

getItem :: Serialize a => JSString -> IO (Either JSString a)

Load a serializable value from local storage. Will fail if the given key does not exist or if the value stored at the key does not match the requested type.

removeItem :: JSString -> IO ()

Remove a value from local storage.