Haste.WebSockets

Description

WebSockets API for Haste.

Documentation

module Haste.Concurrent

data WebSocket a

A WebSocket, parameterized over the type of data that can be sent and received over it.

data WebSocketConfig t

The various event handlers supported by WebSockets.

Constructors

WebSocketConfig

Fields

wsOpenURL :: URL

URL to which to attempt a WebSocket connection.

wsOnClose :: WebSocketCloseEvent -> CIO ()

Called when the connection is closed by the server.

wsOnMessage :: WebSocket t -> t -> CIO ()

Called when a new message arrives.

data WebSocketCloseEvent

Event describing the remote closing of a WebSocket.

Constructors

WebSocketCloseEvent

Fields

wsCloseCode :: Int

Reason for close event, computer-readable version.

wsCloseReason :: JSString

Reason for close event, human-readable version.

wsCloseClean :: Bool

Was the socket closed cleanly?

class WebSocketData a where

Some data that can be sent and received over a WebSocket.

Methods

wsSend :: WebSocket a -> a -> CIO Bool

Send a piece of data to a WebSocket of the appropriate type. Returns False if the data could not be sent, for instance due to a closed socket.

wsOpen :: WebSocketConfig a -> CIO (Maybe (WebSocket a))

Open a new WebSocket connection, returning Just socket when successful.

noHandlers :: WebSocketConfig t

No-op handlers for all WebSocket events.

wsOpenSync :: WebSocketData t => URL -> CIO (Maybe (WebSocket t, CIO (Either WebSocketCloseEvent t)))

Open a WebSocket, returning the socket as well as a blocking CIO computation to await incoming data on the socket. When data arrives on the socket, the waiting computation will return Right received_data. When the socket is closed, the waiting computation will return Left close_event.

wsClose :: WebSocket a -> CIO ()

Close a WebSocket. Closing a WebSocket which has been previously closed is a no-op. Any associated wsOnClose handler will not be called.

withWebSocket

Arguments

:: URL
URL to bind the WebSocket to

-> (WebSocket JSString -> JSString -> CIO ())
Computation to run when new data arrives

-> CIO a
Computation to run when an error occurs

-> (WebSocket JSString -> CIO a)
Computation using the WebSocket

-> CIO a

Run a computation with a WebSocket. The computation will not be executed until a connection to the server has been established. The WebSocket will not be closed after the computation finishes.

withBinaryWebSocket

Arguments

:: URL
URL to bind the WebSocket to

-> (WebSocket Blob -> Blob -> CIO ())
Computation to run when new data arrives

-> CIO a
Computation to run when an error occurs

-> (WebSocket Blob -> CIO a)
Computation using the WebSocket

-> CIO a

Run a computation with a WebSocket. The computation will not be executed until a connection to the server has been established. The WebSocket will not be closed after the computation finishes.

wsSendBlob :: WebSocket Blob -> Blob -> CIO Bool

Send a Blob over a WebSocket.